src/music_theory.py
```python
import numpy as np
from scipy.signal import find_peaks

def applyMusicTheory(dataSet):
    """
    Function to apply principles of music theory to enhance the user experience.
    This function uses the principles of harmony, rhythm, and melody to create a visually stunning and engaging user experience.
    """
    # Convert the data set into a numpy array for easier manipulation
    dataSet = np.array(dataSet)

    # Find the peaks in the data set to represent the rhythm
    peaks, _ = find_peaks(dataSet)

    # Use the peaks to create a melody by mapping the peak values to musical notes
    melody = [mapPeakToNote(peak) for peak in peaks]

    # Use the melody to create a harmony by adding complementary notes
    harmony = [addComplementaryNotes(note) for note in melody]

    # Return the harmony as the final result
    return harmony

def mapPeakToNote(peak):
    """
    Function to map a peak value to a musical note.
    This function uses a simple mapping where the peak value is divided by the maximum possible peak value to get a value between 0 and 1.
    This value is then multiplied by the number of notes in an octave (12) to get the note.
    """
    # Define the maximum possible peak value
    maxPeakValue = np.max(peak)

    # Map the peak value to a note
    note = (peak / maxPeakValue) * 12

    # Return the note
    return note

def addComplementaryNotes(note):
    """
    Function to add complementary notes to a note to create a harmony.
    This function uses the principles of harmony to add notes that complement the original note.
    """
    # Define the complementary notes
    complementaryNotes = [note + interval for interval in [4, 7]]

    # Return the harmony
    return complementaryNotes
```